+++ packages/colony-js-contract-client/src/__tests__/ContractMethod.js
@@ -42,8 +42,12 @@ describe('ContractMethod', () => {
   });
 
   test('Method with input params validates correctly', () => {
-    const input = [['id', 'number']];
-    const inputValues = { id: 1 };
+    const input = [
+      ['taskId', 'number'],
+      ['potId', 'number'],
+      ['domainId', 'number', 1],
+    ];
+    const inputValues = { taskId: 6, potId: 420 };
 
     const method = new ContractMethod({
       client,
@@ -54,25 +58,55 @@ describe('ContractMethod', () => {
     sandbox.spyOn(method, '_validateValue');
 
     expect(method.validate(inputValues)).toBe(true);
-    expect(method._validateValue).toHaveBeenCalledWith(1, 'number', 'id');
+    expect(method._validateValue).toHaveBeenCalledWith(6, input[0]);
+    expect(method._validateValue).toHaveBeenCalledWith(420, input[1]);
 
     // Missing parameters/wrong type
-    [undefined, null, [], 'a', 1].forEach(wrongType => {
+    [undefined, null, [], 'a', 6].forEach(wrongType => {
       expect(() => {
         method.validate(wrongType);
       }).toThrowError('Expected parameters as an object');
     });
 
+    // No parameters
+    expect(() => {
+      method.validate({});
+    }).toThrowError('Missing parameters');
+
+    // Missing parameter
+    expect(() => {
+      method.validate({ taskId: 6 });
+    }).toThrowError('Missing parameters: "potId"');
+
+    // Wine parameters
+    expect(() => {
+      method.validate({
+        caskId: 6,
+        côteId: 420,
+        domaineId: 'Domaine de la Romanee-Conti',
+      });
+    }).toThrowError('Unexpected parameters');
+
     // Extra parameter
     expect(() => {
-      method.validate({ id: 1, somethingElse: 2 });
-    }).toThrowError('Mismatching parameters/method parameters sizes');
+      method.validate({
+        taskId: 6,
+        domainId: 1,
+        potId: 420,
+        somethingElse: 2,
+      });
+    }).toThrowError('Unexpected parameters: "somethingElse"');
+
+    // Extra parameter, without the parameter that has a default value
+    expect(() => {
+      method.validate({ taskId: 6, potId: 420, somethingElse: 1 });
+    }).toThrowError('Unexpected parameters: "somethingElse"');
 
     // Wrong type
     validateValue.mockImplementationOnce(() => false);
     expect(() => {
-      method.validate({ id: 'one' });
-    }).toThrowError('Parameter "id" expected a value of type "number"');
+      method.validate({ taskId: 'six', potId: 420 });
+    }).toThrowError('Parameter "taskId" expected a value of type "number"');
   });
 
   test('Valid values are reported as valid', () => {
@@ -84,7 +118,7 @@ describe('ContractMethod', () => {
       name: 'myMethodName',
     });
     validateValue.mockImplementationOnce(() => true);
-    expect(method._validateValue(1, 'number', 'id')).toBe(true);
+    expect(method._validateValue(1, ['id', 'number'])).toBe(true);
     expect(validateValue).toHaveBeenCalledWith(1, 'number');
   });
 
@@ -100,7 +134,7 @@ describe('ContractMethod', () => {
     // Invalid value
     validateValue.mockImplementationOnce(() => false);
     expect(() => {
-      expect(method._validateValue('abc', 'number', 'id'));
+      expect(method._validateValue('abc', ['id', 'number']));
     }).toThrowError('Parameter "id" expected a value of type "number"');
     expect(validateValue).toHaveBeenCalledWith('abc', 'number');
 
@@ -109,7 +143,7 @@ describe('ContractMethod', () => {
       throw new Error('The reason this validation failed');
     });
     expect(() => {
-      expect(method._validateValue('abc', 'number', 'id'));
+      expect(method._validateValue('abc', ['id', 'number']));
     }).toThrowError(
       'Parameter "id" expected a value of type ' +
         '"number" (The reason this validation failed)',
@@ -120,6 +154,41 @@ describe('ContractMethod', () => {
     expect(() => {
       expect(method._validateValue('abc', 'number', 'id'));
     }).toThrowError('Validation failed for myMethod');
+  });
+
+  test('Valid default values are reported as valid', () => {
+    validateValue.mockImplementationOnce(() => true);
+    const input = [['id', 'number']];
+    const method = new ContractMethod({
+      client,
+      input,
+      name: 'myMethodName',
+    });
+    expect(method._validateValue(undefined, ['id', 'number', 1])).toBe(true);
+    expect(validateValue).toHaveBeenCalledWith(1, 'number');
+  });
+
+  test('Invalid default values are reported as invalid', () => {
+    const input = [['id', 'number']];
+    const method = new ContractMethod({
+      client,
+      input,
+      name: 'myMethodName',
+    });
+    validateValue.mockImplementationOnce(() => false);
+    expect(() => {
+      expect(
+        method._validateValue(undefined, [
+          'id',
+          'number',
+          'a bad default value',
+        ]),
+      );
+    }).toThrowError(
+      'Validation failed for myMethodName: Parameter "id" expected a ' +
+        'value of type "number"',
+    );
+    expect(validateValue).toHaveBeenCalledWith('a bad default value', 'number');
   });
 
   test('Method arguments are processed from input parameters', () => {